+++ js/data-generator.js
@@ -1,77 +1,6 @@
-import { userPhotoMakeID,createdCommentID } from './utitlities-for-generator';
-import { getRandomInteger } from './utilities';
-const ACCEPTABLE_NUMBER_OF_ID = 25;
-
-const VARIANTS_OF_COMMENTS = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают.',
-  'Как можно было поймать такой неудачный момент?!',
-];
-
-const NAMES = [
-  'Алексей',
-  'Мария',
-  'Иван',
-  'Елена',
-  'Дмитрий',
-  'Наталья',
-  'Павел',
-  'Ольга',
-  'Сергей',
-  'Анна',
-  'Владимир',
-  'Татьяна',
-  'Андрей',
-  'Ирина',
-  'Николай',
-  'Екатерина',
-  'Александр',
-  'Марина',
-  'Виктор',
-  'Светлана',
-  'Михаил',
-  'Юлия',
-  'Евгений',
-  'Вероника',
-  'Константин',
-  'Галина',
-  'Антон',
-  'Людмила',
-  'Юрий',
-  'Надежда'
-];
-
-const PHOTOS_DESCRIPTIONS = [
-  'Закат над озером, окрашивающий небо в оранжево-розовые тона.',
-  'Улица старого города, украшенная гирляндами и уличными фонарями.',
-  'Снежный лес и следы животных на свежем снегу.',
-  'Портрет собаки в вязаном свитере, сидящей на скамейке.',
-  'Горная река со скалистыми берегами в осеннем лесу.',
-  'Крупный план капель росы на паутине в лучах утреннего солнца.',
-  'Цветущий сад под голубым облачным небом весной.',
-  'Пейзаж пустыни с каменными формациями на закате.',
-  'Ночное небо, усыпанное звездами, с видимой Млечным Путем.',
-  'Уличный артист, играющий на гитаре перед прохожими.',
-  'Разноцветные воздушные шары, летящие в ясном небе.',
-  'Туман над горной долиной на рассвете.',
-  'Старинный замок, отражающийся в водах озера.',
-  'Макро снимок капель воды на лепестках цветка.',
-  'Праздничный фейерверк над городской набережной.',
-  'Завтрак в постели: круассаны, свежие ягоды и чашка кофе.',
-  'Группа пингвинов на льдине в Антарктике.',
-  'Стая птиц, летящая на фоне закатного неба.',
-  'Зимний пейзаж с заснеженными деревьями и маленьким домиком.',
-  'Дети, играющие с осенними листьями в парке.',
-  'Крупный план цветущего вишневого дерева.',
-  'Силуэты людей, гуляющих по пляжу на закате.',
-  'Городской вид с высотных зданий в сумерки.',
-  'Разноцветные умбрильи, висящие над уличкой.',
-  'Лодки на тихом озере в туманное утро.'
-];
+import { getRandomInteger,userPhotoMakeID,createdCommentID } from './utilities.js';
+import { VARIANTS_OF_COMMENTS,NAMES, PHOTOS_DESCRIPTIONS } from './constans.js';
+export const ACCEPTABLE_NUMBER_OF_ID = 25;
 
 
 const makeComments = () => ({
@@ -82,7 +11,7 @@ const makeComments = () => ({
 });
 
 
-const usersPfotoDescriptions = () => {
+export const usersPfotoDescriptions = () => {
   const newID = userPhotoMakeID.getNewID();
   return {
     id: newID,
@@ -92,5 +21,3 @@ const usersPfotoDescriptions = () => {
     comments: Array.from({length: getRandomInteger(1,4)}, makeComments)
   };
 };
-
-export {usersPfotoDescriptions, ACCEPTABLE_NUMBER_OF_ID};
